requirements.txt
flask
python-escpos